bluetooth: host: conn: Check if *conn is not NULL
This commit adds a warning and a Kconfig option to `bt_conn_le_create`
and `bt_conn_le_create_synced` functions which are meant to warn a user
of a potential leakage of an active connection object.

This change is implemented due to frequent incorrect use of the
connection pointer where a pointer to an existing connection object
was overwritten by `bt_conn_le_create` and `bt_conn_le_create_synced`
functions which lead to sporadic critical bugs.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>