bluetooth: mesh: health_cli: Add callback to pull msg for periodic pub
The Health Client model supports periodic publication and thus requires
the update callback to be initialized.

This commit implements a callback inside the health client model and
adds a wrapper to it so that user can fill up the msg buffer. This way
the Config Server will always accept Config Model Publication Set
message with periodic publication parameters. If the users callback
returns error or not implemented, the periodic publication will be
aborted.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>