bluetooth: host: att: Remove bt_att_create_rsp_pdu
This commit removes `bt_att_create_rsp_pdu` as it is a duplicate of
`bt_att_chan_create_pdu`.

`bt_att_chan_create_pdu` also checks the ATT MTU, so minimal length is added
as argument where it is appropriate. In other places the response
handlers ensures that the ATT MTU check is performed.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>