bluetooth: host: Add a check for num of bt_conn_tx and ACL/ISO bufs
After https://github.com/zephyrproject-rtos/zephyr/pull/72090, each
packet to be sent (either it is ACL or ISO data) has a corresponding
bt_conn_tx object (regardless of whether the callback is used or not).

This means that number of packets Host can send towards Controller is limited
to the smaller of ACL + ISO packets that Controller can receive and number
of bt_conn_tx objects that Host allocated.

The difference between these numbers may result in inefficient resources
usage on either Host's or Controller's side. If Host allocated less
bt_conn_tx objects than number of buffers that Controller allocated for a
certain type of data, the remaining buffers won't be used by Host. And vice
versa, if Host allocated more bt_conn_tx objects than packets that
Controller can consume, the remaining bt_conn_tx objects won't be used
by Host.

This commit adds a check with a warning if the number of bt_conn_tx objects
is not aligned with the ACL/ISO buffers allocated on the controller side
and returned as a result of Le Read Buffer Size v1 or v2 command.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>