bluetooth: host: Prototype removing keys on fly from resolving list
This is a prototype to check if it is possible to remove and re-add keys
to/from Resolving List on fly when conflict happens.

In the prototype, all conflicting keys are removed from the Resolving
List when bonding completes. The keys are re-added back when advertising
starts, and other conflicting keys get removed instead.