bluetooth: host: smp: Don't try to add peer to RL when Privacy disabled
Even if the Privacy feature is disabled, the peer may still send
Identity Information and Identity Address Information commands if it
supports the Privacy feature.

Since we do not support the Privacy feature, there is no need to attempt
adding the peer to the Resolving List. This also eliminates the need to
search for conflicting IRKs, reject the pairing, or remove the
conflicting one.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>