bluetooth: host: Use critical section when making channel ready to send
A public API call that sends ACL data, for example, `bt_gatt_notify` can
be called from a preemptive thread context. This API in turns call
`raise_data_ready` function which adds an L2CAP channel to the
`l2cap_data_ready` list.

The atomic variable used to create a critical section around
`l2cap_data_list` didn't work, because a preemtive thread can be
pre-empted at any time. This means that regardless of `atomic_set`
result, we can't trust it as the thread could be preempted after the
call and the atomic variable state could be changed (specifically by the
TX processor in this case). The same issue with `bt_conn_data_ready`,
which is called next.

When making an L2CAP channel ready for sending data, we need to use a
critical section when appending a channel to the l2cap_data_ready list.
We need to do the same for the `conn_ready` list.

Since cooperative thread can only be rescheduled explicitly, we must
ensure there is no rescheduling points when accessing both, the
`l2cap_data_ready` and `conn_ready` lists.

This change removes the atomic variables used for creating a critical
section for both lists as it doesn't help. Instead,
`k_sched_lock`/`k_sched_unlock` are used where a code can be executed
from a preemtive thread context. `sys_slist_find` is used to check if
a channel or a connection is already in the corresponding list.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>