Bluetooth: Host: Update bt_cmd_send_sync workaround
The workaround in bt_cmd_send_sync is no longer needed when tx_processor
runs on a dedicated bt_taskq and not on system workqueue.

But for defensive programming, we keep the workaround in place and log a
warning if it's triggered. If CONFIG_TEST is enabled, we panic instead.

Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>