DNM: Reproduce deadlock in the event loop
This commit demonstrates a deadlock that occures only with events
on a single-link scenario.

This demonstration relies on auto-initiated procedure, Connection
Paramters Update command and short event pool. The event pool size is
reduced to minimal possible to demonstrate the deadlock in a single-link
scenario.

nrf5340bsim/nrf5340/cpuapp target is used to demonstarte the deadlock.

In case of nrf52_bsim/native (single code device), hci_driver.c
implements a shortcut that bypasses regular flow of events and thus
deadlock never occurs with sync commands.